general debugging, adding features
git-svn-id: https://plugins.svn.wordpress.org/stl-viewer/trunk@1109077 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ stl.php
@@ -143,9 +143,11 @@ if(!class_exists('STLViewer')) {
             }
         }
 		public function deactivate() {
-            $settings = $this->STLViewer_Settings->getSettingsArray();
-            foreach($settings as $setting){
-                delete_option($this->STLViewer_Settings->getSettingPrefix().$setting['name']);
+            if( get_option('stlviewer_delete_settings') == 'checked') {
+                $settings = $this->STLViewer_Settings->getSettingsArray();
+                foreach ($settings as $setting) {
+                    delete_option($this->STLViewer_Settings->getSettingPrefix() . $setting['name']);
+                }
             }
         }
